handling dwnl by layer logic moved to dlg.py file
