+++ bdot10k.py
@@ -134,137 +134,9 @@ class BDOT10k:
         self.dlg = BDOT10kDialogBase()
         self.dlg.show()
 
-    def check_dwnl_path(self, downloadPath):
-        if not downloadPath:
-            QMessageBox.critical(None, "Błąd", "Wskaż lokalizację pobierania.")
-            return False
-        elif not os.path.exists(downloadPath):
-            QMessageBox.critical(None, "Błąd", "Podana lokalizacja nie istnieje.")
-            return False
-        else:
-            return True
-
-    def switch_rbtns_dlgByLayer(self):
-        if self.dlgByLayer.gbOldSchema.isChecked():
-            self.dlgByLayer.rbtnGML.setDisabled(True)
-            self.dlgByLayer.rbtnGPKG.setDisabled(True)
-        else:
-            self.dlgByLayer.rbtnGML.setDisabled(False)
-            self.dlgByLayer.rbtnGPKG.setDisabled(False)
-
     def run_by_layer(self):
-        # if self.first_start == True:
-            # self.first_start = False
         self.dlgByLayer = BDOT10kDialogByLayer()
-
-        # set filters for the map layer combo box - only vector layers
-        self.dlgByLayer.mcbLayer.setFilters(QgsMapLayerProxyModel.PointLayer |
-                                            QgsMapLayerProxyModel.LineLayer |
-                                            QgsMapLayerProxyModel.PolygonLayer)
-
-        self.dlgByLayer.mcbLayer.layerChanged.connect(self.select_by_layer)
-        self.dlgByLayer.btnDwnl.clicked.connect(self.download_by_layer)
-        self.dlgByLayer.gbOldSchema.clicked.connect(self.switch_rbtns_dlgByLayer)
-
-        if self.dlgByLayer.txt:
-            self.dlgByLayer.txt.clear()
-
-        global powiatyTerytByLayer
-        powiatyTerytByLayer = []
-
-        if powiatyTerytByLayer:
-            powiatyTerytByLayer = []
-
-        # show the dialog
         self.dlgByLayer.show()
-
-    def select_by_layer(self):
-        layerForSelection = self.dlgByLayer.mcbLayer.currentLayer()
-        layerPowiatyPath = os.path.join(self.plugin_dir, "powiaty.geojson")
-        layerPowiaty = QgsVectorLayer(layerPowiatyPath, "powiaty", "ogr")
-
-        global powiatyTerytByLayer
-        powiatyTerytByLayer = []
-
-        if self.dlgByLayer.isVisible():
-            if not layerForSelection:
-                QMessageBox.warning(self.dlgByLayer, "Uwaga", "Wybierz warstwę wektorową.")
-            elif layerForSelection and layerForSelection.featureCount() == 0:
-                QMessageBox.warning(self.dlgByLayer, "Uwaga", "Wybrana warstwa nie zawiera obiektów.")
-            else:                
-                if layerPowiaty.crs() != QgsCoordinateReferenceSystem('EPSG:2180'):
-                    layerPowiaty = processing.run("native:reprojectlayer", 
-                        {'INPUT':layerPowiaty,
-                        'TARGET_CRS':QgsCoordinateReferenceSystem('EPSG:2180'),
-                        'OUTPUT':'TEMPORARY_OUTPUT'}
-                    )['OUTPUT']
-
-                powiatySelection = processing.run("native:selectbylocation",
-                    {'INPUT': layerPowiaty,
-                    'PREDICATE': [0],
-                    'INTERSECT': layerForSelection,
-                    'METHOD': 0}
-                )
-
-                powiatySelected = powiatySelection['OUTPUT'].selectedFeatures()
-
-                powiatyTxt = "Powiaty: "
-
-                if powiatySelected:
-                    for feature in powiatySelected:
-                        powiatyTerytByLayer.append(feature["teryt"])
-                        powiatyTxt += feature["teryt"] + " " + feature["nazwa"] + ", "
-
-                    powiatyCount = f"Liczba wyselekcjonowanych powiatów: {len(powiatyTerytByLayer)}"
-                    self.dlgByLayer.txt.clear()
-                    self.dlgByLayer.txt.append(powiatyCount)
-                    self.dlgByLayer.txt.append(powiatyTxt)
-
-                    return powiatyTerytByLayer
-
-                else:
-                    powiatyCount = f"Liczba wyselekcjonowanych powiatów: {len(powiatyTerytByLayer)}"
-                    self.dlgByLayer.txt.clear()
-                    self.dlgByLayer.txt.append(powiatyCount)
-                    QMessageBox.critical(self.dlgByLayer, "Błąd", "Nie znaleziono żadnych powiatów.")
-
-                return powiatyTerytByLayer
-
-        return powiatyTerytByLayer
-
-    def download_by_layer(self):
-        if not powiatyTerytByLayer:
-            QMessageBox.critical(self.dlgByLayer, "Błąd", "Brak powiatów do pobrania.")
-        else:
-            downloadPath = self.dlgByLayer.dwnlPath.filePath()
-
-            if self.dlgByLayer.gbOldSchema.isChecked():
-                oldSchema = True
-                if self.dlgByLayer.rbtnSHPold.isChecked():
-                    bdot10kDataFormat = 'SHP'
-                elif self.dlgByLayer.rbtnGMLold.isChecked():
-                    bdot10kDataFormat = 'GML'
-            else:
-                oldSchema = False
-                if self.dlgByLayer.rbtnGML.isChecked():
-                    bdot10kDataFormat = 'GML'
-                elif self.dlgByLayer.rbtnGPKG.isChecked():
-                    bdot10kDataFormat = 'GPKG'
-
-            if self.check_dwnl_path(downloadPath):
-                QgsMessageLog.logMessage(f'Lokalizacja pobierania: {downloadPath}', 'BDOT10k', level=Qgis.MessageLevel.Info)
-                QgsMessageLog.logMessage('Pobieranie paczek dla powiatów: ' + str(sorted(powiatyTerytByLayer)), 'BDOT10k', level=Qgis.MessageLevel.Info)
-
-                task = DownloadBdotTask(
-                    description="Pobieranie paczek BDOT10k",
-                    downloadPath=downloadPath,
-                    oldSchema=oldSchema,
-                    bdot10kDataFormat=bdot10kDataFormat,
-                    powiatyTerytList=powiatyTerytByLayer,
-                    iface=self.iface
-                )
-
-                self.taskManager.addTask(task)
 
     def info(self):
         self.dlgInfo = BDOT10kDialogInfo()
